fix: extract date calculations to avoid config import in tests
The calculateReviewDates function was in createTrialReviewThread.ts which
imports config.ts, causing CI failures when env vars aren't set. Moved to
a standalone dateCalculations.ts with no config dependency.
